requirements.txt
Flask
google-generativeai
numpy
openpyxl
pandas
pillow
PyMuPDF
python-dotenv
sentence-transformers